S41.cpp

+++ S41.cpp
@@ -1,20 +1,20 @@
 #include <iostream>
+
 using namespace std;
 
-int squared_sum(double a, double b)
-{
+double squared_sum(double a, double b) {
     double c;
-           c = (a + b)*(a+b);
+    c = (a + b) * (a + b);
     return c;
 }
-int main()
-{
+
+int main() {
     setlocale(LC_ALL, "Russian");
-       double a, b;
-cout << "Введите первое число: ";
-          cin >> a;
-cout << "Введите второе число: ";
-          cin >> b;
-cout << "Квадрат суммы чисел равен: " << squared_sum(a, b);
-return 0;
+    double a, b;
+    cout << "Введите первое число: ";
+    cin >> a;
+    cout << "Введите второе число: ";
+    cin >> b;
+    cout << "Квадрат суммы чисел равен: " << squared_sum(a, b);
+    return 0;
 }
